change floats to doubles
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@2834 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/routers/router.hpp
@@ -54,8 +54,8 @@ protected:
   int _input_speedup;
   int _output_speedup;
   
-  float _internal_speedup;
-  float _partial_internal_cycles;
+  double _internal_speedup;
+  double _partial_internal_cycles;
 
   int _crossbar_delay;
   int _credit_delay;
